convert javascript to model for game
